fixed the code
